feat: support custom pks in contract:deploy
